media: at91sam9x5-video: modify videodev "vfl_dir" policy
With the new ioctl functions a v4l2 device can not be a receiver and a transmitter.

This is not compatible with at91sam9x5-video overlay driver.
This driver uses ioctl on "V4L2_BUF_TYPE_VIDEO_OVERLAY" (receiver mode) and
"V4L2_BUF_TYPE_VIDEO_OUTPUT" (transmitter mode).
This patch adds the capabilities for a video4linux device to act as a receiver
and a transmitter.

Signed-off-by: Patrice Vilchez <patrice.vilchez@atmel.com>