atmel_nand: NFC: should handle multiple IRQ
This patch fix the errors happen when NFC data transfer:
  atmel_nand 80000000.nand: Time out to wait for interrupt: 0x00010000
  atmel_nand 80000000.nand: something wrong, No XFR_DONE interrupt comes.

The root cause is in current interrupt handling, we read the ISR but
only handle one interrupt. If there are more than one interrupt comes.
The other interrupt will be lost as read the ISR will clear the ISR.

So when we enable transfer data in NFC command, we may receive
NFC_CMD_DONE and NFC_XFR_DONE interrupts in same time.

This patch enable multiple IRQ handling. When we enable transfer data in
NFC command, we ask to wait two interrupts NFC_CMD_DONE and
NFC_XFR_DONE.

Signed-off-by: Josh Wu <josh.wu@atmel.com>

Conflicts:

	drivers/mtd/nand/atmel_nand.c

Signed-off-by: Josh Wu <josh.wu@atmel.com>