ARM: at91/aic: use real number of connected IRQ lines
Instead of relying on the maximum number of IRQ lines that the AIC can handle,
give real number of IRQ lines for each SoC. Note that some AIC doesn't allow to
use a non-connected peripheral ID. It is actually the reason why we implement
this modification.

By using a different compatibility string, we can describe the real hardware
configuration. This variable in the driver is updated in a wrapper function as
we stick to the standard definition of declaring a DT IRQ chip.

Signed-off-by: Nicolas Ferre <nicolas.ferre@atmel.com>