mtd: atmel_nand: nfc: reduce the range of NFC command address for iomap
This can fix following error:
  vmap allocation for size 268439552 failed: use vmalloc=<size> to increase size.
  atmel_nand_nfc 90000000.nfc: ioremap failed for resource [mem 0x90000000-0x9fffffff]
  atmel_nand_nfc: probe of 90000000.nfc failed with error -12

In the NFCDATA_STATUS, the NFCBUSY is the highest bit, the 27th bit.
And this bit will not combined with other bits. So we can remap this
region with small space: 4bytes, which can use for readl().
So that we can save 0x08000000 virtual address space.

Add the NFCBUSY address to map the resource #1. Then move the other
resource to next number.

Meantime, we need modify the sama5d3/4.dtsi to add busy command address
in nfc node.

Signed-off-by: Josh Wu <josh.wu@atmel.com>