ARM: at91: dts: sama5d3/sama5d4: reduce NFC command registers memory region
To get NFC status, we need to read the NFC command registers, which is a
256M memory mapping address. To check the status, you just read such
address from NFC command registers.
For example, to check NFCBUSY (bit 27), you need to read offset 0x08000000
(bit 27 is set to 1) of the NFC command register.

Currently, the atmel nand driver use HSMC_SR to check NFCBUSY, so don't need
to read offset 0x08000000 for check NFCBUSY status.

Since the MSB is bit 26, we only need to map 0x0~0x08000000 for NFC
command registers.

Signed-off-by: Josh Wu <josh.wu@atmel.com>