ARM: at91: sama5d3/dts: enable led d3 only for boards where isi is disabled
As led d3 is conflict with camera sensor reset pin PE24, and only some of
rev.C boards not enable isi with camera sensors. For those boards we
will enable d3.

So in summary, only enable led d3 for:
  sama5d31_revc.dts
  sama5d33_revc.dts
  sama5d34_revc.dts

In the meantime, I sorted the dts file by alphabet order.

Signed-off-by: Josh Wu <josh.wu@atmel.com>