ARM: at91/dt: remove *_hdmi.dtsi in dts files
As the driver is absolutely needed for the whole LCD subsystem to probe
correctly, I remove the *_hdmi.dtsi from all .dts files.
It is too dangerous to risk the LCD not to be probed just because you forget to
select one driver in your config file.

Just make sure to select the sil902x driver when you uncomment the _hdmi.dtsi
include and you'll be ready to use it.

Signed-off-by: Nicolas Ferre <nicolas.ferre@atmel.com>