pinctrl: at91-pio4: add debounce configuration
Add debounce configuration without taking care of the value requested
because we can't have a different debounce period for lines inside the
same group. Moreover the register to configure this value is a secure
register even if the pin is set as non secure.

The debouncing filter can filter a pulse with a duration of less than
1/2 programmable divided slow clock period:
tdiv_slck = (( DIV + 1) x 2) x tslck, DIV is set to 0 by default and
won't be changed.

Signed-off-by: Ludovic Desroches <ludovic.desroches@atmel.com>