v4l2: clk: convert dt node to i2c device if the of named v4l2_clk is not find
The soc-camera host will be probed and register a v4l2_clk, but on the
moment, if i2c device is not avaiable, so the v4l2_clk's name is used
the of string.

But the i2c sensor is probed, the will search the v4l2_clk with the i2c
device name, like 1-0030.

So that would never find the clock since the soc-camera host register
the clock with name of dt node string.

The solution is, if we cannot find the clock, we try to convert the dt
node to i2c device, then compare it. If matches then return the clock.

Signed-off-by: Josh Wu <josh.wu@atmel.com>