media: atmel-isc: add gammar correction to make rgb565 work correct
As if we enable CFA, it will output RGB valid data to GAM module.
In GAM module, it will do a 12 to 10 bits compression, that means 2 bits
information will be lost if you don't use gammar table.
If we set up a gammar table then 12 to 10 bits compression work
correctly.

So this patch will enable gammar table to make it Bayer -> RGB565 work
correctly.

Signed-off-by: Josh Wu <josh.wu@atmel.com>